requirements.txt
Flask
Flask-Login
mysql-connector-python
qrcode
pillow